src/requirements.txt
arrow==1.2.3
lxml==4.9.3
PyYAML==6.0
requests==2.31.0
tinydb==4.8.0